# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Yunora AI Question Generation Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chapters:
    get:
      operationId: listChapters
      tags: [chapters]
      summary: List chapters
      security:
        - bearerAuth: []
      parameters:
        - name: subjectId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterList"
    post:
      operationId: createChapter
      tags: [chapters]
      summary: Create a chapter
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /chapters/{id}:
    patch:
      operationId: updateChapter
      tags: [chapters]
      summary: Update chapter
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
    delete:
      operationId: deleteChapter
      tags: [chapters]
      summary: Delete chapter
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── TOPICS ───────────────────────────────────────────────────────────────────
...
  /generation/start:
    post:
      operationId: startGeneration
      tags: [generation]
      summary: Start AI question generation job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerationRequest"
      responses:
        "202":
          description: Job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJob"
  /generation/jobs:
    get:
      operationId: listGenerationJobs
      tags: [generation]
      summary: List generation jobs
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Generation jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJobList"
  /generation/jobs/{jobId}:
    get:
      operationId: getGenerationJob
      tags: [generation]
      summary: Get generation job status and results
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJobDetail"

  # ── AI PROVIDERS ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Chapter:
      type: object
      required: [id, name, subjectId, orderIndex]
      properties:
        id: { type: integer }
        name: { type: string }
        orderIndex: { type: integer }
        subjectId: { type: integer }
        subjectName: { type: ["string", "null"] }
        isActive: { type: boolean }
        topicsCount: { type: integer }
        questionsCount: { type: integer }
        createdAt: { type: string, format: date-time }
    ChapterInput:
      type: object
      required: [name, subjectId, orderIndex]
      properties:
        name: { type: string }
        orderIndex: { type: integer }
        subjectId: { type: integer }
        isActive: { type: boolean, default: true }
    ChapterUpdate:
      type: object
      properties:
        name: { type: string }
        orderIndex: { type: integer }
        isActive: { type: boolean }
    ChapterList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── TOPICS ────────────────────────────────────────────────────────────────
...
    Question:
      type: object
      required: [id, question, questionType, difficulty, difficultyScore, correctAnswer, explanation, qualityScore]
      properties:
        id: { type: integer }
        question: { type: string }
        questionType: { type: string }
        difficulty: { type: string, enum: [easy, medium, hard] }
        difficultyScore: { type: integer }
        correctAnswer: { type: string }
        options: { type: ["string", "null"] }
        explanation: { type: string }
        qualityScore: { type: number }
        estimatedSolveTime: { type: integer }
        learningObjective: { type: ["string", "null"] }
        topicId: { type: ["integer", "null"] }
        topicName: { type: ["string", "null"] }
        chapterId: { type: ["integer", "null"] }
        chapterName: { type: ["string", "null"] }
        subjectId: { type: ["integer", "null"] }
        subjectName: { type: ["string", "null"] }
        boardId: { type: ["integer", "null"] }
        boardName: { type: ["string", "null"] }
        standardId: { type: ["integer", "null"] }
        standardName: { type: ["string", "null"] }
        modelUsed: { type: ["string", "null"] }
        providerId: { type: ["integer", "null"] }
        generatedAt: { type: string, format: date-time }
...
    GenerationRequest:
      type: object
      required: [boardId, standardId, subjectId, chapterId, topicId, questionType, difficulty, count, providerId, model]
      properties:
        boardId: { type: integer }
        standardId: { type: integer }
        subjectId: { type: integer }
        chapterId: { type: integer }
        topicId: { type: integer }
        questionType: { type: string }
        difficulty: { type: string, enum: [easy, medium, hard] }
        count: { type: integer, minimum: 1, maximum: 50 }
        providerId: { type: integer }
        model: { type: string }
    GenerationJob:
      type: object
      required: [jobId, status, createdAt]
      properties:
        jobId: { type: string }
        status: { type: string, enum: [pending, processing, completed, failed] }
        totalRequested: { type: integer }
        createdAt: { type: string, format: date-time }
    GenerationJobList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/GenerationJobSummary"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    GenerationJobSummary:
      type: object
      required: [jobId, status, totalRequested, createdAt]
      properties:
        jobId: { type: string }
        status: { type: string, enum: [pending, processing, completed, failed] }
        totalRequested: { type: integer }
        totalGenerated: { type: ["integer", "null"] }
        topicName: { type: ["string", "null"] }
        subjectName: { type: ["string", "null"] }
        difficulty: { type: string }
        questionType: { type: string }
        model: { type: string }
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"] }
    GenerationJobDetail:
      type: object
      required: [jobId, status, totalRequested, createdAt]
      properties:
        jobId: { type: string }
        status: { type: string, enum: [pending, processing, completed, failed] }
        totalRequested: { type: integer }
        totalGenerated: { type: ["integer", "null"] }
        agentLogs: { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"] }

    # ── AI PROVIDERS ──────────────────────────────────────────────────────────
...