# lib/api-zod/src/generated/types/chapter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yunora AI Question Generation Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface Chapter {
  id: number;
  name: string;
  orderIndex: number;
  subjectId: number;
  /** @nullable */
  subjectName?: string | null;
  isActive?: boolean;
  topicsCount?: number;
  questionsCount?: number;
  createdAt?: Date;
}