# lib/api-zod/src/generated/types/chapterUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yunora AI Question Generation Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface ChapterUpdate {
  name?: string;
  orderIndex?: number;
  isActive?: boolean;
}